Use GitHub Actions for CI
This sets up a GitHub action to run the gem's tests when a pull request
is created and on pushes to the main branch.

Tests are run in a test matrix that tests ruby versions 3.1 and 3.2.
Should we want to test against the development release of ruby (head),
this also includes the ability to allow that to fail without CI failing.
That will give us the information that this currently fails, should head
be released, without needing to take action, because it is not released.

However, head is not currently part of the test matrix.

Redis is set up according to GitHub's [redis documentation](https://docs.github.com/en/actions/using-containerized-services/creating-redis-service-containers).